Setup Basic DC SEA and Josh Mehling's SEA Models
The basic model uses an armature controlled DC motor, and does position
control. The PID is not tuned at all, and the motor properties are all
just one right now. We can find some example problem with real values or
find some motor specs if we want to test the setup.

Josh's model was in his paper, and the output of the system is already
designed for impedance control, so if we want to switch control schemes,
we'll need to change the transfer functions a little bit.